Add change username to setings page && made toggle theme fancier
