+++ frontend/src/routes.ts
@@ -5,7 +5,6 @@ import Login from "./routes/auth/Login.svelte";
 import Callback from "./routes/auth/Callback.svelte";
 import Authenticate2FA from "./routes/auth/TwoFactorAuth.svelte";
 import Manage2FA from "./routes/settings/Manage2FA.svelte";
-import ManageUserName from "./routes/settings/ManageUserName.svelte";
 import Settings from "./routes/settings/Settings.svelte";
 import JoinChannel from "./routes/chat/JoinChannel.svelte";
 
@@ -15,7 +14,6 @@ export default {
   "/settings": Settings,
   "/chat": Chat,
   "/settings/2fa": Manage2FA,
-  "/settings/username": ManageUserName,
   "/auth/login": Login,
   "/auth/oauth2callback": Callback,
   "/auth/2fa": Authenticate2FA,
